Moved Get DB Connection method to utils
