+++ AzureFunctions/km-rag-function/function_app.py
@@ -1,10 +1,10 @@
 import azure.functions as func
-from azure.identity import DefaultAzureCredential
 import logging
 import openai
 from azurefunctions.extensions.http.fastapi import Request, StreamingResponse
 import asyncio
 import os
+from utils import dbutils
 
 from typing import Annotated
 
@@ -19,8 +19,6 @@ from semantic_kernel.core_plugins.time_plugin import TimePlugin
 from semantic_kernel.functions.kernel_arguments import KernelArguments
 from semantic_kernel.functions.kernel_function_decorator import kernel_function
 from semantic_kernel.kernel import Kernel
-import pyodbc
-import struct
 
 # from semantic_kernel import Kernel
 # from semantic_kernel.connectors.ai.open_ai import AzureChatCompletion
@@ -39,45 +37,6 @@ app = func.FunctionApp(http_auth_level=func.AuthLevel.ANONYMOUS)
 
 # search_endpoint = os.environ.get("AZURE_AI_SEARCH_ENDPOINT") 
 # search_key = os.environ.get("AZURE_AI_SEARCH_API_KEY")
-
-# get database connection
-def get_db_connection():
-    driver = "{ODBC Driver 17 for SQL Server}"
-    server = os.environ.get("SQLDB_SERVER")
-    database = os.environ.get("SQLDB_DATABASE")
-    username = os.environ.get("SQLDB_USERNAME")
-    password = os.environ.get("SQLDB_PASSWORD")
-
-    # Attempt connection using Username & Password
-    try:
-        credential = DefaultAzureCredential()
-
-        token_bytes = credential.get_token(
-            "https://database.windows.net/.default"
-        ).token.encode("utf-16-LE")
-        token_struct = struct.pack(f"<I{len(token_bytes)}s", len(token_bytes), token_bytes)
-        SQL_COPT_SS_ACCESS_TOKEN = (
-            1256  # This connection option is defined by microsoft in msodbcsql.h
-        )
-
-        # Set up the connection
-        connection_string = f"DRIVER={driver};SERVER={server};DATABASE={database};"
-        conn = pyodbc.connect(
-            connection_string, attrs_before={SQL_COPT_SS_ACCESS_TOKEN: token_struct}
-        )
-
-        logging.info("Connected using Default Azure Credential")
-
-        return conn
-    except pyodbc.Error as e:
-        logging.error(f"Failed with Default Credential: {str(e)}")
-        conn = pyodbc.connect(
-            f"DRIVER={driver};SERVER={server};DATABASE={database};UID={username};PWD={password}",
-            timeout=5
-        )
-        
-        logging.info("Connected using Username & Password")
-        return conn
 
 class ChatWithDataPlugin:
     @kernel_function(name="Greeting", description="Respond to any greeting or general questions")
@@ -154,8 +113,7 @@ class ChatWithDataPlugin:
             sql_query = sql_query.replace("```sql",'').replace("```",'')
             #print(sql_query)
 
-            conn = get_db_connection()
-            # conn = pyodbc.connect(connectionString)
+            conn = dbutils.get_db_connection()
             cursor = conn.cursor()
             cursor.execute(sql_query)
             answer = ''
